// src/proto/xudrpc_pb.d.ts
    export type AsObject = {
    }
// ...
export class GetInfoRequest extends jspb.Message { 

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetInfoRequest.AsObject;
    static toObject(includeInstance: boolean, msg: GetInfoRequest): GetInfoRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetInfoRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetInfoRequest;
    static deserializeBinaryFromReader(message: GetInfoRequest, reader: jspb.BinaryReader): GetInfoRequest;
}

export namespace GetInfoRequest {
    export type AsObject = {
    }
}

export class GetInfoResponse extends jspb.Message { 
    getVersion(): string;
    setVersion(value: string): GetInfoResponse;

    getNodePubKey(): string;
    setNodePubKey(value: string): GetInfoResponse;

    clearUrisList(): void;
    getUrisList(): Array<string>;
    setUrisList(value: Array<string>): GetInfoResponse;
    addUris(value: string, index?: number): string;

    getNumPeers(): number;
    setNumPeers(value: number): GetInfoResponse;

    getNumPairs(): number;
    setNumPairs(value: number): GetInfoResponse;


    hasOrders(): boolean;
    clearOrders(): void;
    getOrders(): OrdersCount | undefined;
    setOrders(value?: OrdersCount): GetInfoResponse;


    getLndMap(): jspb.Map<string, LndInfo>;
    clearLndMap(): void;


    hasRaiden(): boolean;
    clearRaiden(): void;
    getRaiden(): RaidenInfo | undefined;
    setRaiden(value?: RaidenInfo): GetInfoResponse;

    getAlias(): string;
    setAlias(value: string): GetInfoResponse;

    getNetwork(): string;
    setNetwork(value: string): GetInfoResponse;

    clearPendingSwapHashesList(): void;
    getPendingSwapHashesList(): Array<string>;
    setPendingSwapHashesList(value: Array<string>): GetInfoResponse;
    addPendingSwapHashes(value: string, index?: number): string;


    hasConnext(): boolean;
    clearConnext(): void;
    getConnext(): ConnextInfo | undefined;
    setConnext(value?: ConnextInfo): GetInfoResponse;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GetInfoResponse.AsObject;
    static toObject(includeInstance: boolean, msg: GetInfoResponse): GetInfoResponse.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GetInfoResponse, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GetInfoResponse;
    static deserializeBinaryFromReader(message: GetInfoResponse, reader: jspb.BinaryReader): GetInfoResponse;
}

export namespace GetInfoResponse {
    export type AsObject = {
        version: string,
        nodePubKey: string,
        urisList: Array<string>,
        numPeers: number,
        numPairs: number,
        orders?: OrdersCount.AsObject,

        lndMap: Array<[string, LndInfo.AsObject]>,
        raiden?: RaidenInfo.AsObject,
        alias: string,
        network: string,
        pendingSwapHashesList: Array<string>,
        connext?: ConnextInfo.AsObject,
    }
}
// ...
export class ListOrdersRequest extends jspb.Message { 
    getPairId(): string;
    setPairId(value: string): ListOrdersRequest;

    getOwner(): ListOrdersRequest.Owner;
    setOwner(value: ListOrdersRequest.Owner): ListOrdersRequest;

    getLimit(): number;
    setLimit(value: number): ListOrdersRequest;

    getIncludeAliases(): boolean;
    setIncludeAliases(value: boolean): ListOrdersRequest;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ListOrdersRequest.AsObject;
    static toObject(includeInstance: boolean, msg: ListOrdersRequest): ListOrdersRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ListOrdersRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ListOrdersRequest;
    static deserializeBinaryFromReader(message: ListOrdersRequest, reader: jspb.BinaryReader): ListOrdersRequest;
}

export namespace ListOrdersRequest {
    export type AsObject = {
        pairId: string,
        owner: ListOrdersRequest.Owner,
        limit: number,
        includeAliases: boolean,
    }

    export enum Owner {
    BOTH = 0,
    OWN = 1,
    PEER = 2,
    }

}

export class ListOrdersResponse extends jspb.Message { 

    getOrdersMap(): jspb.Map<string, Orders>;
    clearOrdersMap(): void;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ListOrdersResponse.AsObject;
    static toObject(includeInstance: boolean, msg: ListOrdersResponse): ListOrdersResponse.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ListOrdersResponse, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ListOrdersResponse;
    static deserializeBinaryFromReader(message: ListOrdersResponse, reader: jspb.BinaryReader): ListOrdersResponse;
}

export namespace ListOrdersResponse {
    export type AsObject = {

        ordersMap: Array<[string, Orders.AsObject]>,
    }
}

export class ListPairsRequest extends jspb.Message { 

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ListPairsRequest.AsObject;
    static toObject(includeInstance: boolean, msg: ListPairsRequest): ListPairsRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ListPairsRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ListPairsRequest;
    static deserializeBinaryFromReader(message: ListPairsRequest, reader: jspb.BinaryReader): ListPairsRequest;
}

export namespace ListPairsRequest {
    export type AsObject = {
    }
}

export class ListPairsResponse extends jspb.Message { 
    clearPairsList(): void;
    getPairsList(): Array<string>;
    setPairsList(value: Array<string>): ListPairsResponse;
    addPairs(value: string, index?: number): string;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ListPairsResponse.AsObject;
    static toObject(includeInstance: boolean, msg: ListPairsResponse): ListPairsResponse.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ListPairsResponse, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ListPairsResponse;
    static deserializeBinaryFromReader(message: ListPairsResponse, reader: jspb.BinaryReader): ListPairsResponse;
}

export namespace ListPairsResponse {
    export type AsObject = {
        pairsList: Array<string>,
    }
}
// ...
export namespace LndInfo {
    export type AsObject = {
        status: string,
        channels?: Channels.AsObject,
        chainsList: Array<Chain.AsObject>,
        blockheight: number,
        urisList: Array<string>,
        version: string,
        alias: string,
    }
}
// ...
export class Order extends jspb.Message { 
    getPrice(): number;
    setPrice(value: number): Order;

    getQuantity(): number;
    setQuantity(value: number): Order;

    getPairId(): string;
    setPairId(value: string): Order;

    getId(): string;
    setId(value: string): Order;


    hasNodeIdentifier(): boolean;
    clearNodeIdentifier(): void;
    getNodeIdentifier(): NodeIdentifier | undefined;
    setNodeIdentifier(value?: NodeIdentifier): Order;

    getLocalId(): string;
    setLocalId(value: string): Order;

    getCreatedAt(): number;
    setCreatedAt(value: number): Order;

    getSide(): OrderSide;
    setSide(value: OrderSide): Order;

    getIsOwnOrder(): boolean;
    setIsOwnOrder(value: boolean): Order;

    getHold(): number;
    setHold(value: number): Order;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Order.AsObject;
    static toObject(includeInstance: boolean, msg: Order): Order.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Order, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Order;
    static deserializeBinaryFromReader(message: Order, reader: jspb.BinaryReader): Order;
}

export namespace Order {
    export type AsObject = {
        price: number,
        quantity: number,
        pairId: string,
        id: string,
        nodeIdentifier?: NodeIdentifier.AsObject,
        localId: string,
        createdAt: number,
        side: OrderSide,
        isOwnOrder: boolean,
        hold: number,
    }
}
// ...
export namespace Orders {
    export type AsObject = {
        buyOrdersList: Array<Order.AsObject>,
        sellOrdersList: Array<Order.AsObject>,
    }
}
// ...
export namespace OrdersCount {
    export type AsObject = {
        peer: number,
        own: number,
    }
}

export class OrderUpdate extends jspb.Message { 

    hasOrder(): boolean;
    clearOrder(): void;
    getOrder(): Order | undefined;
    setOrder(value?: Order): OrderUpdate;


    hasOrderRemoval(): boolean;
    clearOrderRemoval(): void;
    getOrderRemoval(): OrderRemoval | undefined;
    setOrderRemoval(value?: OrderRemoval): OrderUpdate;


    getOrderUpdateCase(): OrderUpdate.OrderUpdateCase;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): OrderUpdate.AsObject;
    static toObject(includeInstance: boolean, msg: OrderUpdate): OrderUpdate.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: OrderUpdate, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): OrderUpdate;
    static deserializeBinaryFromReader(message: OrderUpdate, reader: jspb.BinaryReader): OrderUpdate;
}

export namespace OrderUpdate {
    export type AsObject = {
        order?: Order.AsObject,
        orderRemoval?: OrderRemoval.AsObject,
    }

    export enum OrderUpdateCase {
        ORDER_UPDATE_NOT_SET = 0,
    
    ORDER = 1,

    ORDER_REMOVAL = 2,

    }

}
// ...
export namespace RaidenInfo {
    export type AsObject = {
        status: string,
        address: string,
        channels?: Channels.AsObject,
        version: string,
        chain: string,
    }
}
// ...
export class SubscribeOrdersRequest extends jspb.Message { 
    getExisting(): boolean;
    setExisting(value: boolean): SubscribeOrdersRequest;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): SubscribeOrdersRequest.AsObject;
    static toObject(includeInstance: boolean, msg: SubscribeOrdersRequest): SubscribeOrdersRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: SubscribeOrdersRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): SubscribeOrdersRequest;
    static deserializeBinaryFromReader(message: SubscribeOrdersRequest, reader: jspb.BinaryReader): SubscribeOrdersRequest;
}

export namespace SubscribeOrdersRequest {
    export type AsObject = {
        existing: boolean,
    }
}